+++ apps/bot/src/domains/resource/errors.ts
@@ -1,0 +1,10 @@
+import type { ResourceName } from '@one-piece/db';
+
+import { AppError } from '../../discord/errors.js';
+
+export class InsufficientResourceError extends AppError {
+  constructor(resourceName: ResourceName) {
+    super(`Ressource insuffisante : ${resourceName}.`, 'warn');
+    this.name = 'InsufficientResourceError';
+  }
+}
