+++ apps/bot/src/domains/resource/service.ts
@@ -1,7 +1,6 @@
 import type { DbOrTransaction, ResourceName } from '@one-piece/db';
 
-import { ValidationError } from '../../discord/errors.js';
-
+import { InsufficientResourceError } from './errors.js';
 import * as resourceRepository from './repository.js';
 
 type ResourceCost = {
@@ -10,10 +9,11 @@ type ResourceCost = {
 };
 
 export async function debitResourcesByName(playerId: number, resources: Array<ResourceCost>, client: DbOrTransaction): Promise<void> {
+  // TODO: Optimiser ça en batch.
   for (const { name, quantity } of resources) {
     const debited = await resourceRepository.debitResourceByName(playerId, name, quantity, client);
     if (!debited) {
-      throw new ValidationError(`Ressource insuffisante : ${name}.`);
+      throw new InsufficientResourceError(name);
     }
   }
 }
