+++ apps/bot/src/domains/ship/utils/assert-player-owns-ship.ts
@@ -1,9 +1,9 @@
-import { ValidationError } from '../../../discord/errors.js';
+import { ForbiddenError } from '../../../discord/errors.js';
 import * as playerRepository from '../../player/repository.js';
 
 export async function assertPlayerOwnsShip(playerId: number, discordId: string): Promise<void> {
   const player = await playerRepository.findByIdOrThrow(playerId);
   if (player.discordId === discordId) return;
 
-  throw new ValidationError("Tu ne peux pas améliorer le navire de quelqu'un d'autre.");
+  throw new ForbiddenError("Tu ne peux pas améliorer le navire de quelqu'un d'autre.");
 }
